Prevent host reallocation for active reservations
Extending the length of a physical host reservation can lead Blazar to
reallocate the reservation to new hosts if some of the reserved hosts
are not available during the extended lease period. This can happen even
if the lease has already started, as long as there is no instance
running on any of the hosts to reallocate.

This can cause a race condition: see upstream change
I62a62cb2497b2c22de722ae71bd060f2f3935010.

In Chameleon, it is not a good user experience to have already started
lease change hosts, which users might not notice.